requirements.txt
pyserial>=3.5
fastapi>=0.104
uvicorn[standard]>=0.24
